requirements.txt
python-gnupg==0.5.2
boto3>=1.34.0
flask>=3.0.0
flask-swagger-ui>=4.11.1
